+++ sdk/eventhub/azure-mgmt-eventhub/generated_samples/eh_alias_list.py
@@ -38,6 +38,6 @@ def main():
         print(item)
 
 
-# x-ms-original-file: 2026-01-01/disasterRecoveryConfigs/EHAliasList.json
+# x-ms-original-file: 2026-07-01-preview/disasterRecoveryConfigs/EHAliasList.json
 if __name__ == "__main__":
     main()