+++ sdk/eventhub/azure-mgmt-eventhub/generated_samples/eh_consumer_group_get.py
@@ -39,6 +39,6 @@ def main():
     print(response)
 
 
-# x-ms-original-file: 2026-01-01/ConsumerGroup/EHConsumerGroupGet.json
+# x-ms-original-file: 2026-07-01-preview/ConsumerGroup/EHConsumerGroupGet.json
 if __name__ == "__main__":
     main()