+++ sdk/eventhub/azure-mgmt-eventhub/generated_samples/eh_event_hub_authorization_rule_delete.py
@@ -38,6 +38,6 @@ def main():
     )
 
 
-# x-ms-original-file: 2026-01-01/EventHubs/EHEventHubAuthorizationRuleDelete.json
+# x-ms-original-file: 2026-07-01-preview/EventHubs/EHEventHubAuthorizationRuleDelete.json
 if __name__ == "__main__":
     main()