+++ sdk/eventhub/azure-mgmt-eventhub/generated_samples/eh_event_hub_create.py
@@ -64,6 +64,6 @@ def main():
     print(response)
 
 
-# x-ms-original-file: 2026-01-01/EventHubs/EHEventHubCreate.json
+# x-ms-original-file: 2026-07-01-preview/EventHubs/EHEventHubCreate.json
 if __name__ == "__main__":
     main()