+++ sdk/eventhub/azure-mgmt-eventhub/CHANGELOG.md
@@ -1,5 +1,55 @@
 # Release History
 
+## 12.0.0b3 (2026-07-22)
+
+### Features Added
+
+  - Client `EventHubManagementClient` added operation group `fabric_shortcuts`
+  - Client `EventHubManagementClient` added operation group `upgrade_preferences_operations`
+  - Added model `ExceptionWindow`
+  - Added enum `ExceptionWindowAction`
+  - Added model `FabricShortcut`
+  - Added model `FabricShortcutConfiguration`
+  - Added model `FabricShortcutProperties`
+  - Added enum `FabricShortcutStatus`
+  - Added enum `FabricShortcutType`
+  - Added model `MaintenanceWindow`
+  - Added enum `UpgradePreferenceDayOfWeek`
+  - Added model `UpgradePreferences`
+  - Added model `UpgradePreferencesProperties`
+  - Added model `UpgradeStatus`
+  - Added operation group `FabricShortcutsOperations`
+  - Added operation group `UpgradePreferencesOperationsOperations`
+
+### Breaking Changes
+
+  - Deleted or renamed model `AccessKeys`
+  - Deleted or renamed model `AvailableCluster`
+  - Deleted or renamed model `AvailableClustersList`
+  - Deleted or renamed model `CheckNameAvailabilityResult`
+  - Deleted or renamed model `EHNamespaceIdContainer`
+  - Deleted or renamed model `EHNamespaceIdListResult`
+  - Deleted or renamed model `ErrorAdditionalInfo`
+  - Deleted or renamed model `ErrorDetail`
+  - Deleted or renamed model `ErrorResponse`
+  - Deleted or renamed model `NetworkRuleSetListResult`
+  - Deleted or renamed model `NetworkSecurityPerimeter`
+  - Deleted or renamed model `NetworkSecurityPerimeterConfiguration`
+  - Deleted or renamed model `NetworkSecurityPerimeterConfigurationList`
+  - Deleted or renamed model `NetworkSecurityPerimeterConfigurationProperties`
+  - Deleted or renamed model `NetworkSecurityPerimeterConfigurationPropertiesProfile`
+  - Deleted or renamed model `NetworkSecurityPerimeterConfigurationPropertiesResourceAssociation`
+  - Deleted or renamed model `NspAccessRule`
+  - Deleted or renamed model `NspAccessRuleProperties`
+  - Deleted or renamed model `NspAccessRulePropertiesSubscriptionsItem`
+  - Deleted or renamed model `Operation`
+  - Deleted or renamed model `OperationDisplay`
+  - Deleted or renamed model `PrivateLinkResource`
+  - Deleted or renamed model `PrivateLinkResourceProperties`
+  - Deleted or renamed model `PrivateLinkResourcesListResult`
+  - Deleted or renamed model `ProvisioningIssue`
+  - Deleted or renamed model `ProvisioningIssueProperties`
+
 ## 12.0.0b2 (2026-07-07)
 
 ### Features Added